add SSH transport method
Add support for SSH transport in virt-v2v. The installation role was
extended to create new SSH key (for vdsm user) if there is no key yet.
Wrapper was extended to support the method.

A technical requirement from QEMU and virt-v2v is that ssh-agent is
running and contains the key to be used. Wrapper spawns the agent,
adds all keys from ~/.ssh, runs virt-v2v and kills the agent after
conversion.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>