wrapper: fix running as vdsm user
It is again possible to run as vdsm user. Rather cumbersome method has
to be used to make export domain still supported option. It is needed to
read the input data, decide whether we can run as vdsm and then feed it
again to the newly spawned instance.

When using the export domain the wrapper still runs as root. This is
necessary to be able to mount the NFS share.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>