wrapper: automatically find ISO with drivers
This adds new key `install_drivers` that enables installation of Windows
drivers. If `virtio_win` is not specified then autodetection is
attempted. The ISO domain is scanned and filenames are matched against
several patterns with varying priority. The best match is selected.

`install_drivers` defaults to false, however for backward compatibility,
when `install_drivers` is not present but `virtio_win` is present, it is
assumed that install_drivers is true.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>