wrapper: pass insecure connection option to virt-v2v
The commit 9e934ec798f0ac9c19dc added option to allow insecure
connections in oVirt SDK. We can further pass the option to virt-v2v
using '-oo rhv-verifypeer' for the same reason.

At the moment the default in virt-v2v is 'false', i.e. to perform
insecure connections. By passing the option we cannot make the
connections less secure, we can only strengthen the security.

Even though it may not appear so, this change is backward compatible. We
already do calls to oVirt SDK in the wrapper. So in the environment with
broken SSL the 'insecure_connection' has to be already set. If it is not
set, it means there is proper SSL setup and enabling verification in
virt-v2v means no harm.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>