spec: install playbooks into secondary location
Playbooks are installed into the documents directory (under examples/).
Unfortunately the path contains the version component in it. This makes
it difficult to use the playbooks for other parties if they just want to
us them as-is.

To ease the situation for third parties playbooks are now also provided
in auxilary directory.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>