wrapper: pass --root=first to virt-v2v
In case there are multiple boot disks in the VM, virt-v2v stops and asks
which one to use. We have no way of knowing which is the right one to
pick, but we definitely don't want to be blocked by this. It seems that
the best thing we can do is to pick the first one and see what happens.

Note that virt-v2v does not support conversion of multi-boot VMs.

Bug-Url: https://bugzilla.redhat.com/1590618
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>