wrapper: re-initialize cURL after fork
Re-initialize cURL. This is necessary to re-initialze the PKCS #11
security tokens in NSS. Because as the documentation says:

    It is an error to try to use a PKCS#11 crypto module in a process
    before it has been initialized in that process, even if the module
    was initialized in the parent process.

Should we fail to do so any use of SDK after the fork() would lead to
the cURL error:

    A PKCS #11 module returned CKR_DEVICE_ERROR, indicating that a
    problem has occurred with the token or slot.

Other alternative would be to disable the fork check by using
NSS_STRICT_NOFORK environment variable[1]. Probably this should be also
safe in our case because the parent process exits immediately and there
is no risk of reusing same tokens in two processes. But doing the full
reset is much better.

[1] https://developer.mozilla.org/en-US/docs/Mozilla/Projects/NSS/Reference/NSS_environment_variables

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>