roles: make sure installed packages are at latest version
Instead of just checkinf if packages are installed we update to the
latest version. This is done for all packages except the ones used to
build the VDDK plugin. For those we still require just their presence.

To override this and return to previous behavior user can set variable
`v2v_yum_check: present`.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>