roles: make nbdkit checking code more compatible
The commit 0d05a8a69d introduced checking for nbdkit version.
Unfortunately the code requires relatively new version of python-jinja2
(2.10). Such version is not yet available on EL. Code was rewritten to
not reqiuire the new features. It is ugglier than original solution, but
works on older versions and is in fact faster than the pure Ansible
solution.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>