Add garbage collector for the V2VVMware objects
if "cnv.io/temporary = true" label is detected, the GC sets obj.spec.timeToLive to "creationTimestamp + DELAY".
Once the timeToLive is reached, the object is removed.

The DELAY is set to 2 hours.

With this design, the actual deletion time can be postponed if needed (not recently the case).

Note:
The standard k8s deletetionTimestamp can not be used as it is not supported for anything else than Pods.
In addition, it would not support prolonging the time, once set.