ansible: don't remove wrapper package on oVirt hosts
The wrapper package is installed as dependency of
ovirt-host-dependencies now and removing it means removing also bunch of
other RPMs. And we don't want that so drop the removal step from
uninstall role.

We used to remove the wrapper mainly to avoid leaving untracked files on
the host. This is not really a problem now that the wrapper is in its
own RPM.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>