First attempt at modularising the codebase
This is just a code movement with relative imports (which we can now properly
use thanks to using setuptools).  There are few things that could benefit from
this, one example could be restricting access to `hard_error()` only to places
where it can be called, but that's just an idea for another day.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>