wrapper: duplicate logs only for Kubevirt
The fact that wrapper runs in container or not daemonized does not mean
we should duplicate the logs. This behavior makes sense most sense only
for Kubevirt where it is more natural to monitor just POD output and
where it's not possible to access the logs in the container later. For
other environments (UCI) it is expected that logs are stored in attached
volume and thus persisted. So keep the duplication only for Kubevirt.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>