wrapper: systemd: clean failed units at exit (RHBZ#1749347)
When systemd unit fails it is left in the list of units so that
administrator can inspect the error. This is unfavorable because the
generated names are not guaranteed to be unique and name duplicates can
appear. In case of duplicate name the service fails to start before the
previous failed instance is removed.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>